+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Implement comments using Prettier's API.
 - Add support for `prettier-ignore` directive ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/3)).
+- Use internal HTTP server instead of Nailgun for built in parser.
 - Fix unstable comments in between If/Else blocks.
 - Fix unstable comments in between Try/Catch/Finally blocks.
 - Fix unstable comments in WhereCompoundExpr.